requirements.txt
PyMuPDF
pydantic
streamlit
python-docx
python-dotenv
google-generativeai